only pass PrivilegedMode to baggageclaim for linux workers
PrivilegedMode is only valid for Linux workers using containerd
