+++ worker/baggageclaim/baggageclaimcmd/command_test.go
@@ -18,15 +18,17 @@ var _ = Describe("Baggage Claim Command Runner", func() {
 			var cmd baggageclaimcmd.BaggageclaimCommand
 			cmd.DisableUserNamespaces = true
 			It("selects the NoopNamespacer", func() {
-				priv, unpriv := cmd.SelectNamespacers(logger, bespec.FullPrivilegedMode)
+				priv, unpriv := cmd.SelectNamespacers(logger)
 				Expect(priv).To(Equal(uidgid.NoopNamespacer{}))
 				Expect(unpriv).To(Equal(uidgid.NoopNamespacer{}))
 			})
 		})
 		Context("when in full privilege mode", func() {
 			var cmd baggageclaimcmd.BaggageclaimCommand
+			cmd.PrivilegedMode = bespec.FullPrivilegedMode
+
 			It("selects separate privileged and unprivileged namespacers", func() {
-				priv, unpriv := cmd.SelectNamespacers(logger, bespec.FullPrivilegedMode)
+				priv, unpriv := cmd.SelectNamespacers(logger)
 				var testCmd exec.Cmd
 				priv.NamespaceCommand(&testCmd)
 				Expect(testCmd.SysProcAttr.UidMappings[0].HostID).To(Equal(0))
@@ -36,8 +38,10 @@ var _ = Describe("Baggage Claim Command Runner", func() {
 		})
 		Context("when in FUSE-only privilege mode", func() {
 			var cmd baggageclaimcmd.BaggageclaimCommand
+			cmd.PrivilegedMode = bespec.FUSEOnlyPrivilegedMode
+
 			It("selects the unprivileged namespacer for both privilege levels", func() {
-				priv, unpriv := cmd.SelectNamespacers(logger, bespec.FUSEOnlyPrivilegedMode)
+				priv, unpriv := cmd.SelectNamespacers(logger)
 				var testCmd exec.Cmd
 				priv.NamespaceCommand(&testCmd)
 				Expect(testCmd.SysProcAttr.UidMappings[0].HostID).To(Not(Equal(0)))
@@ -47,8 +51,10 @@ var _ = Describe("Baggage Claim Command Runner", func() {
 		})
 		Context("when in ignore privilege mode", func() {
 			var cmd baggageclaimcmd.BaggageclaimCommand
+			cmd.PrivilegedMode = bespec.IgnorePrivilegedMode
+
 			It("selects the unprivileged namespacer for both privilege levels", func() {
-				priv, unpriv := cmd.SelectNamespacers(logger, bespec.IgnorePrivilegedMode)
+				priv, unpriv := cmd.SelectNamespacers(logger)
 				var testCmd exec.Cmd
 				priv.NamespaceCommand(&testCmd)
 				Expect(testCmd.SysProcAttr.UidMappings[0].HostID).To(Not(Equal(0)))
