+++ worker/baggageclaim/cmd/baggageclaim/main.go
@@ -5,7 +5,6 @@ import (
 	"os"
 
 	"github.com/concourse/concourse/worker/baggageclaim/baggageclaimcmd"
-	"github.com/concourse/concourse/worker/runtime/spec"
 	"github.com/jessevdk/go-flags"
 )
 
@@ -20,7 +19,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	err = cmd.Execute(args, spec.FullPrivilegedMode)
+	err = cmd.Execute(args)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
